move `replace-permalink.js` into `utils/`
- removed unnecessary use of lodash
