+++ packages/url-utils/lib/utils/index.js
@@ -4,5 +4,6 @@ module.exports = {
     deduplicateSubdirectory: require('./deduplicate-subdirectory'),
     isSSL: require('./is-ssl'),
     relativeToAbsolute: require('./relative-to-absolute'),
+    replacePermalink: require('./replace-permalink'),
     stripSubdirectoryFromPath: require('./strip-subdirectory-from-path')
 };
