+++ packages/url-utils/test/unit/url-utils.test.js
@@ -3,7 +3,6 @@
 require('../utils');
 
 const sinon = require('sinon');
-const moment = require('moment-timezone');
 const UrlUtils = require('../../lib/index');
 
 const constants = {
@@ -746,111 +745,19 @@ describe('UrlUtils', function () {
     });
 
     describe('replacePermalink', function () {
-        it('permalink is /:slug/, timezone is default', function () {
-            const testData = {
-                slug: 'short-and-sweet'
-            };
-            const postLink = '/short-and-sweet/';
+        it('calls outs to utils/replace-permalink', function () {
+            const utils = new UrlUtils();
 
-            (new UrlUtils()).replacePermalink('/:slug/', testData).should.equal(postLink);
-        });
+            const spy = sinon.spy();
+            utils._utils.replacePermalink = spy;
 
-        it('permalink is /:year/:month/:day/:slug/, blog timezone is Los Angeles', function () {
-            const testData = {
-                slug: 'short-and-sweet',
-                published_at: new Date('2016-05-18T06:30:00.000Z')
-            };
-            const timezone = 'America/Los_Angeles';
-            const postLink = '/2016/05/17/short-and-sweet/';
+            utils.replacePermalink('testPermalink', 'testResource', 'testTimezone');
 
-            (new UrlUtils()).replacePermalink('/:year/:month/:day/:slug/', testData, timezone).should.equal(postLink);
-        });
-
-        it('permalink is /:year/:month/:day/:slug/, blog timezone is Asia Tokyo', function () {
-            const testData = {
-                slug: 'short-and-sweet',
-                published_at: new Date('2016-05-18T06:30:00.000Z')
-            };
-            const timezone = 'Asia/Tokyo';
-            const postLink = '/2016/05/18/short-and-sweet/';
-
-            (new UrlUtils()).replacePermalink('/:year/:month/:day/:slug/', testData, timezone).should.equal(postLink);
-        });
-
-        it('permalink is /:year/:id/:author/, TZ is LA', function () {
-            const testData = {
-                id: 3,
-                primary_author: {slug: 'joe-blog'},
-                slug: 'short-and-sweet',
-                published_at: new Date('2016-01-01T00:00:00.000Z')
-            };
-            const timezone = 'America/Los_Angeles';
-            const postLink = '/2015/3/joe-blog/';
-
-            (new UrlUtils()).replacePermalink('/:year/:id/:author/', testData, timezone).should.equal(postLink);
-        });
-
-        it('permalink is /:year/:id:/:author/, TZ is Berlin', function () {
-            const testData = {
-                id: 3,
-                primary_author: {slug: 'joe-blog'},
-                slug: 'short-and-sweet',
-                published_at: new Date('2016-01-01T00:00:00.000Z')
-            };
-            const timezone = 'Europe/Berlin';
-            const postLink = '/2016/3/joe-blog/';
-
-            (new UrlUtils()).replacePermalink('/:year/:id/:author/', testData, timezone).should.equal(postLink);
-        });
-
-        it('permalink is /:primary_tag/:slug/ and there is a primary_tag', function () {
-            const testData = {
-                slug: 'short-and-sweet',
-                primary_tag: {slug: 'bitcoin'}
-            };
-            const timezone = 'Europe/Berlin';
-            const postLink = '/bitcoin/short-and-sweet/';
-
-            (new UrlUtils()).replacePermalink('/:primary_tag/:slug/', testData, timezone).should.equal(postLink);
-        });
-
-        it('permalink is /:primary_tag/:slug/ and there is NO primary_tag', function () {
-            const testData = {
-                slug: 'short-and-sweet'
-            };
-            const timezone = 'Europe/Berlin';
-            const postLink = '/all/short-and-sweet/';
-
-            (new UrlUtils()).replacePermalink('/:primary_tag/:slug/', testData, timezone).should.equal(postLink);
-        });
-
-        it('shows "undefined" for unknown route segments', function () {
-            const testData = {
-                slug: 'short-and-sweet'
-            };
-            const timezone = 'Europe/Berlin';
-            const postLink = '/undefined/short-and-sweet/';
-
-            (new UrlUtils()).replacePermalink('/:tag/:slug/', testData, timezone).should.equal(postLink);
-        });
-
-        it('post is not published yet', function () {
-            const testData = {
-                id: 3,
-                slug: 'short-and-sweet',
-                published_at: null
-            };
-            const timezone = 'Europe/London';
-
-            const nowMoment = moment().tz('Europe/London');
-
-            let postLink = '/YYYY/MM/DD/short-and-sweet/';
-
-            postLink = postLink.replace('YYYY', nowMoment.format('YYYY'));
-            postLink = postLink.replace('MM', nowMoment.format('MM'));
-            postLink = postLink.replace('DD', nowMoment.format('DD'));
-
-            (new UrlUtils()).replacePermalink('/:year/:month/:day/:slug/', testData, timezone).should.equal(postLink);
+            const {calledOnce, firstCall} = spy;
+            calledOnce.should.be.true('called once');
+            firstCall.args[0].should.eql('testPermalink');
+            firstCall.args[1].should.eql('testResource');
+            firstCall.args[2].should.eql('testTimezone');
         });
     });
